% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accpal}
\alias{accpal}
\title{a palette with 7 elements, derived from RColorBrewer 'accent'}
\format{vector}
\usage{
accpal
}
\description{
a palette with 7 elements, derived from RColorBrewer 'accent'
}
\keyword{datasets}
