% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cell_type_3.5k}
\alias{cell_type_3.5k}
\title{the vector of cell types assigned for 3584 cells from Darmanis 2017}
\format{vector}
\usage{
cell_type_3.5k
}
\description{
the vector of cell types assigned for 3584 cells from Darmanis 2017
}
\keyword{datasets}
