% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{litrse}
\alias{litrse}
\title{sample of data on chr1 from Darmanis 2017/conquer quantifications}
\format{RangedSummarizedExperiment}
\usage{
litrse
}
\description{
sample of data on chr1 from Darmanis 2017/conquer quantifications
}
\keyword{datasets}
