% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_exprs.R
\name{smooth_exprs}
\alias{smooth_exprs}
\title{smooth_exprs will compute an 'RNA-seq based CNV profile" as
in Darmanis et al. Cell 2017}
\usage{
smooth_exprs(rse, seqname = "1", id = 1, bandwidth = 0.1,
  tx = function(x) log(x + 1))
}
\arguments{
\item{rse}{a RangedSummarizedExperiment instance}

\item{seqname}{character(1)}

\item{id}{sample identifier or index to process}

\item{bandwidth}{a fraction defaulting to 0.1 that will
determine window width}

\item{tx}{a function defaulting to function(x)log(x+1)}
}
\value{
a data.frame with x=smoothed position and y=smoothed tx(expression)
}
\description{
smooth_exprs will compute an 'RNA-seq based CNV profile" as
in Darmanis et al. Cell 2017
}
