% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spin.R
\name{spin_tsne}
\alias{spin_tsne}
\title{demonstrate 3d t-SNE reduction}
\usage{
spin_tsne()
}
\description{
demonstrate 3d t-SNE reduction
}
\note{
This is a result of Rtsne with dims=3 and no other tuning, applied to the 739 x 3584 Darmanis 2017 subset
}
