% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{t1.3d}
\alias{t1.3d}
\title{output of default Rtsne run with dims=3 on 739 gene Darmanis data}
\format{list}
\usage{
t1.3d
}
\description{
output of default Rtsne run with dims=3 on 739 gene Darmanis data
}
\keyword{datasets}
